% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{calculate_importance_ci}
\alias{calculate_importance_ci}
\title{Calculate Variable Importance with Bootstrap Confidence Intervals}
\usage{
calculate_importance_ci(cv_models, verbose = TRUE)
}
\arguments{
\item{cv_models}{List of CV model objects}

\item{verbose}{Print progress}
}
\value{
List with importance table and plot
}
\description{
Calculate Variable Importance with Bootstrap Confidence Intervals
}
\keyword{internal}
