% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{create_spatial_weights}
\alias{create_spatial_weights}
\title{Create Spatial Weights Matrix}
\usage{
create_spatial_weights(data, coords = NULL, k_neighbors = 20)
}
\arguments{
\item{data}{Spatial data}

\item{coords}{Optional coordinate matrix}

\item{k_neighbors}{Number of neighbors}
}
\value{
Spatial weights listw object
}
\description{
Create Spatial Weights Matrix
}
\keyword{internal}
