% Generated by roxygen2: do not edit by hand
\name{plot.SArf}
\alias{plot.SArf}
\title{Plot method for SArf objects}
\usage{
\method{plot}{SArf}(x, type = "all", ...)
}
\arguments{
\item{x}{An object of class 'SArf'}

\item{type}{Type of plot: "all" (default), "moran", "importance", "ale", "map"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Creates diagnostic plots from SArf analysis
}
\examples{
\dontrun{
results <- SArf(outcome ~ x1 + x2, data = mydata)
plot(results)  # All plots
plot(results, type = "importance")  # Just importance plot
plot(results, type = "ale")  # Just ALE plots
}
}
