% Generated by roxygen2: do not edit by hand
\name{print.SArf}
\alias{print.SArf}
\title{Print method for SArf objects}
\usage{
\method{print}{SArf}(x, ...)
}
\arguments{
\item{x}{An object of class 'SArf'}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for SArf objects
}
