% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cv.R
\name{spatial_cv_rf}
\alias{spatial_cv_rf}
\title{Spatial Cross-Validation Random Forest with Within-Fold Spatial Lag}
\usage{
spatial_cv_rf(
  formula,
  data,
  spatial_weights,
  n_folds = 5,
  n_bootstrap = 20,
  block_range = 1000,
  num_trees = 500,
  mtry = NULL,
  seed = 1111,
  verbose = TRUE
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Spatial data}

\item{spatial_weights}{Spatial weights listw object}

\item{n_folds}{Number of CV folds}

\item{n_bootstrap}{Number of bootstrap iterations}

\item{block_range}{Spatial blocking range in meters}

\item{num_trees}{Number of trees in RF}

\item{mtry}{Number of variables to split on}

\item{seed}{Random seed}

\item{verbose}{Print progress}
}
\value{
List with CV results
}
\description{
Spatial Cross-Validation Random Forest with Within-Fold Spatial Lag
}
\keyword{internal}
