% Generated by roxygen2: do not edit by hand
\name{summary.SArf}
\alias{summary.SArf}
\title{Summary method for SArf objects}
\usage{
\method{summary}{SArf}(object, ...)
}
\arguments{
\item{object}{An object of class 'SArf'}

\item{...}{Additional arguments (not used)}
}
\description{
Provides detailed summary of SArf analysis results
}
