% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{test_morans_i}
\alias{test_morans_i}
\title{Test for Spatial Autocorrelation using Moran's I}
\usage{
test_morans_i(
  data,
  variable,
  coords = NULL,
  k_neighbors = 20,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{data}{Spatial data (sf object or data.frame)}

\item{variable}{Name of variable to test}

\item{coords}{Optional coordinate matrix}

\item{k_neighbors}{Number of neighbors for spatial weights}

\item{alpha}{Significance level}

\item{verbose}{Print progress messages}
}
\value{
List with test results and plot
}
\description{
Test for Spatial Autocorrelation using Moran's I
}
\keyword{internal}
