% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SArf.R
\name{plot.SArf}
\alias{plot.SArf}
\title{Plot method for SArf objects}
\usage{
\method{plot}{SArf}(x, which = "all", ...)
}
\arguments{
\item{x}{An object of class 'SArf'}

\item{which}{Character: which plot to display? Options: "all", "moran", 
"importance", "ale", or "comparison" (default: "all")}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for SArf objects
}
