% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SArf.R
\name{summary.SArf}
\alias{summary.SArf}
\title{Summary method for SArf objects}
\usage{
\method{summary}{SArf}(object, ...)
}
\arguments{
\item{object}{An object of class 'SArf'}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for SArf objects
}
